Occurrences of ammonoid fossils

Many ammonoid (Cephalopoda, Mollusca) fossil occurrences from different moments of the geological past. Much information (i.e., extra columns) was removed from the original dataset to make it more compact, but it can be fully accessed by the data URL. This dataset is part of the package and is licensed under the Creative Commons Attribution 4.0 International License (CC BY 4.0).

Usage

ammonoidea_fossil

Format

A data.frame with 58111 rows and 13 columns.

phylum

Organism phylum

class

Organism taxonomic class

order

Organism taxonomic order

family

Organism taxonomic family

genus

Organism genus

species

Organism specific name

early_interval

Earlier known geological period of occurrence

late_interval

Later known geological period of occurrence

max_ma

Occurrence's oldest time boundary in million years

min_ma

Occurrence's newest time boundary in million years

midpoint

Midpoint between max_ma and min_ma

lng

Longitude of place where occurrence was found. Follows decimal degree format.

lat

Latitude of place where occurrence was found. Follows decimal degree format.

Source

The Paleobiology Database (downloaded on 2022-03-11).
Data URL: http://paleobiodb.org/data1.2/occs/list.csv?datainfo&rowcount&base_name=Ammonoidea&show=full,classext,genus,subgenus,acconly,ident,img,etbasis,strat,lith,env,timebins,timecompare,resgroup,ref,ent,entname,crmod